Move rev not msg type to separate module
