+++ aries_vcx/src/protocols/revocation_notification/receiver/state_machine.rs
@@ -1,6 +1,6 @@
 use messages::ack::please_ack::AckOn;
-use messages::issuance::revocation_ack::RevocationAck;
-use messages::issuance::revocation_notification::{RevocationNotification, RevocationFormat};
+use messages::revocation_notification::revocation_ack::RevocationAck;
+use messages::revocation_notification::revocation_notification::{RevocationNotification, RevocationFormat};
 
 use crate::error::prelude::*;
 use crate::protocols::SendClosure;
