+++ aries_vcx/src/protocols/revocation_notification/receiver/states/received.rs
@@ -1,4 +1,4 @@
-use messages::issuance::revocation_notification::RevocationNotification;
+use messages::revocation_notification::revocation_notification::RevocationNotification;
 
 #[derive(Serialize, Deserialize, Debug, Clone, PartialEq)]
 pub struct NotificationReceivedState {
