+++ messages/src/a2a/mod.rs
@@ -11,8 +11,8 @@ use crate::connection::request::Request;
 use crate::connection::response::SignedResponse;
 use crate::discovery::disclose::Disclose;
 use crate::discovery::query::Query;
-use crate::issuance::revocation_ack::RevocationAck;
-use crate::issuance::revocation_notification::RevocationNotification;
+use crate::revocation_notification::revocation_ack::RevocationAck;
+use crate::revocation_notification::revocation_notification::RevocationNotification;
 use crate::problem_report::ProblemReport as CommonProblemReport;
 use crate::forward::Forward;
 use crate::issuance::credential::Credential;
