+++ messages/src/revocation_notification/mod.rs
@@ -1,0 +1,2 @@
+pub mod revocation_notification;
+pub mod revocation_ack;
