+++ messages/src/revocation_notification/revocation_ack.rs
@@ -1,4 +1,3 @@
-// TODO: Move to a separate module
 use crate::ack::Ack;
 
 pub type RevocationAck = Ack;
